Base class added
