+++ orangeHrmDemo/src/test/java/orangeHrmTestcases/TC_003_LogOut.java
@@ -15,25 +15,7 @@ import org.openqa.selenium.firefox.FirefoxDriver;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 
-public class TC_003_LogOut {
-
-	WebDriver driver;
-
-	@BeforeClass
-	@Parameters("crossbrowser")
-	public void beforeClass(String crossbrowser) {
-		// launch browser
-		if (crossbrowser.equals("chrome")) {
-			System.setProperty("webdriver.chrome.driver", "./drivers\\chromedriver.exe");
-			driver = new ChromeDriver();
-		} else if (crossbrowser.equals("firefox")) {
-			System.setProperty("webdriver.gecko.driver", "./drivers\\geckodriver.exe");
-			driver = new FirefoxDriver();
-		}
-		
-		driver.manage().window().maximize();
-		driver.manage().timeouts().implicitlyWait(20, TimeUnit.SECONDS);
-	}
+public class TC_003_LogOut extends BaseClass{
 
 	@Test
 	public void OrangeHrm_TC003_logOut() throws InterruptedException  {
@@ -54,9 +36,8 @@ public class TC_003_LogOut {
 		Assert.assertTrue(login.isLogoutSuccessful());
 	}
 
-	@AfterClass
-	public void afterClass() {
-		driver.quit();
-	}
+	/*
+	 * @AfterClass public void afterClass() { driver.quit(); }
+	 */
 
 }
