+++ orangeHrmDemo/src/test/java/orangeHrmTestcases/TC_007_Candidatecalendar.java
@@ -16,27 +16,9 @@ import org.openqa.selenium.firefox.FirefoxDriver;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 
-public class TC_007_Candidatecalendar {
-	WebDriver driver;
+public class TC_007_Candidatecalendar extends BaseClass{
 	
-	 @BeforeClass
-	 @Parameters("crossbrowser")
-	 public void beforeClass(String crossbrowser) {
-		
-		// launch browser
-				if(crossbrowser.equals("chrome")) {
-					System.setProperty("webdriver.chrome.driver","./drivers\\chromedriver.exe");  
-					driver = new ChromeDriver();
-					}else if (crossbrowser.equals("firefox")) {
-						System.setProperty("webdriver.gecko.driver", "./drivers\\geckodriver.exe");
-						driver = new FirefoxDriver();
-					}
-		 
-				driver.manage().window().maximize();
-				driver.manage().timeouts().implicitlyWait(20, TimeUnit.SECONDS);
-			 }
-	
-	 @Test
+ @Test
   public void TC_008_candidateCalendar() {
 		
 		LoginPageObject login = new LoginPageObject(driver); 
@@ -60,9 +42,8 @@ public class TC_007_Candidatecalendar {
 	 }
   
 
-  @AfterClass
-  public void afterClass() {
-	 driver.close();
-  }
+		/*
+		 * @AfterClass public void afterClass() { driver.close(); }
+		 */
 
 }
