+++ orangeHrmDemo/test-output/old/OrangeHRM automation regression/classes.html
@@ -18,7 +18,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -56,7 +56,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -86,7 +86,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -116,7 +116,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -146,7 +146,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -176,7 +176,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -206,7 +206,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
@@ -236,7 +236,7 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>beforeClass</td>
+<td>launchBrowser</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeMethod</td>
